ci: add expect pass and expect fail test scripts
To test _printf, we need the ability to assert it behaves as expected.
We need to make this as easy as possible; ideally, we automate it.
Otherwise, we won't stay on top of checking _printf after each change.

This commit adds two scripts that check the behaviour of _printf.
'expect_pass' asserts that a program will output something.
'expect_fail' asserts that a program will NOT output something.

Usage notes are in each script file. Subsequent commits will introduce
a build system that uses these scripts to automate all checks of our
test entry points. Test entry points contain a 'main' function and are
stored in the directory '/tests'. Such a setup takes our repository
closer to a continuous integration and continuous delivery model.